+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Add dj_escorted_off / dj_booed_off events
 * Add fan_added / fan_removed events
 * Add user_name_updated / user_avatar_updated events